+++ src/gui.rs
@@ -20,12 +20,12 @@ use serde::{Deserialize, Serialize};
 use serialport::{DataBits, FlowControl, Parity, StopBits};
 
 use crate::color_picker::{color_picker_widget, color_picker_window, COLORS};
+use crate::custom_highlighter::highlight_impl;
 use crate::data::{DataContainer, SerialDirection};
 use crate::serial::{clear_serial_settings, save_serial_settings, Device, SerialDevices};
 use crate::toggle::toggle;
 use crate::FileOptions;
 use crate::{APP_INFO, PREFS_KEY};
-use crate::custom_highlighter::highlight_impl;
 
 const DEFAULT_FONT_ID: FontId = FontId::new(14.0, FontFamily::Monospace);
 pub const RIGHT_PANEL_WIDTH: f32 = 350.0;
@@ -45,91 +45,12 @@ const SAVE_PLOT_SHORTCUT: KeyboardShortcut = KeyboardShortcut::new(
 const CLEAR_PLOT_SHORTCUT: KeyboardShortcut =
     KeyboardShortcut::new(egui::Modifiers::COMMAND, egui::Key::X);
 
-#[derive(Clone)]
-#[allow(unused)]
-pub enum Print {
-    Empty,
-    Message(String),
-    Error(String),
-    Debug(String),
-    Ok(String),
-}
-
 #[derive(PartialEq)]
 pub enum WindowFeedback {
     None,
     Waiting,
     Clear,
     Cancel,
-}
-
-impl Print {
-    pub fn scroll_area_message(
-        &self,
-        gui_conf: &GuiSettingsContainer,
-    ) -> Option<ScrollAreaMessage> {
-        match self {
-            Print::Empty => None,
-            Print::Message(s) => {
-                let color = if gui_conf.dark_mode {
-                    Color32::WHITE
-                } else {
-                    Color32::BLACK
-                };
-                Some(ScrollAreaMessage {
-                    label: "[MSG] ".to_owned(),
-                    content: s.to_owned(),
-                    color,
-                })
-            }
-            Print::Error(s) => {
-                let color = Color32::RED;
-                Some(ScrollAreaMessage {
-                    label: "[ERR] ".to_owned(),
-                    content: s.to_owned(),
-                    color,
-                })
-            }
-            Print::Debug(s) => {
-                let color = if gui_conf.dark_mode {
-                    Color32::YELLOW
-                } else {
-                    Color32::LIGHT_RED
-                };
-                Some(ScrollAreaMessage {
-                    label: "[DBG] ".to_owned(),
-                    content: s.to_owned(),
-                    color,
-                })
-            }
-            Print::Ok(s) => {
-                let color = Color32::GREEN;
-                Some(ScrollAreaMessage {
-                    label: "[OK] ".to_owned(),
-                    content: s.to_owned(),
-                    color,
-                })
-            }
-        }
-    }
-}
-
-#[allow(dead_code)]
-pub struct ScrollAreaMessage {
-    label: String,
-    content: String,
-    color: Color32,
-}
-
-pub fn print_to_console(print_lock: &Arc<RwLock<Vec<Print>>>, message: Print) {
-    match print_lock.write() {
-        Ok(mut write_guard) => {
-            write_guard.push(message);
-        }
-        Err(e) => {
-            println!("Error while writing to print_lock: {}", e);
-        }
-    }
 }
 
 #[derive(Serialize, Deserialize, PartialEq, Debug, Clone)]
@@ -184,12 +105,10 @@ pub struct MyApp {
     serial_devices: SerialDevices,
     plotting_range: usize,
     plot_serial_display_ratio: f32,
-    console: Vec<Print>,
     picked_path: PathBuf,
     plot_location: Option<egui::Rect>,
     data: DataContainer,
     gui_conf: GuiSettingsContainer,
-    print_lock: Arc<RwLock<Vec<Print>>>,
     device_lock: Arc<RwLock<Device>>,
     devices_lock: Arc<RwLock<Vec<String>>>,
     connected_lock: Arc<RwLock<bool>>,
@@ -215,7 +134,6 @@ pub struct MyApp {
 #[allow(clippy::too_many_arguments)]
 impl MyApp {
     pub fn new(
-        print_lock: Arc<RwLock<Vec<Print>>>,
         data_lock: Arc<RwLock<DataContainer>>,
         device_lock: Arc<RwLock<Device>>,
         devices_lock: Arc<RwLock<Vec<String>>>,
@@ -232,15 +150,11 @@ impl MyApp {
             device: "".to_string(),
             old_device: "".to_string(),
             data: DataContainer::default(),
-            console: vec![Print::Message(
-                "waiting for serial connection..,".to_owned(),
-            )],
             connected_lock,
             device_lock,
             devices_lock,
             device_idx: 0,
             serial_devices: devices,
-            print_lock,
             gui_conf,
             data_lock,
             save_tx,
@@ -249,8 +163,8 @@ impl MyApp {
             plotting_range: usize::MAX,
             plot_serial_display_ratio: 0.45,
             command: "".to_string(),
-            show_sent_cmds: false,
-            show_timestamps: false,
+            show_sent_cmds: true,
+            show_timestamps: true,
             save_raw: false,
             eol: "\\r\\n".to_string(),
             colors: vec![COLORS[0]],
@@ -323,18 +237,18 @@ impl MyApp {
             // Height
             let top_spacing = 5.0;
             let panel_height = ui.available_size().y;
-            let mut plot_height:f32 = 0.0;
+            let mut plot_height: f32 = 0.0;
 
             if self.serial_devices.number_of_plots[self.device_idx] > 0 {
-                let height:f32 ;
-                height = ui.available_size().y * self.plot_serial_display_ratio;
+                let height = ui.available_size().y * self.plot_serial_display_ratio;
                 plot_height = height;
                 // need to subtract 12.0, this seems to be the height of the separator of two adjacent plots
-                plot_height =
-                plot_height / (self.serial_devices.number_of_plots[self.device_idx] as f32) - 12.0;
+                plot_height = plot_height
+                    / (self.serial_devices.number_of_plots[self.device_idx] as f32)
+                    - 12.0;
             }
 
-            let mut plot_ui_heigh:f32 = 0.0;
+            let mut plot_ui_heigh: f32 = 0.0;
 
             ui.add_space(top_spacing);
             ui.horizontal(|ui| {
@@ -344,7 +258,6 @@ impl MyApp {
                         self.data = read_guard.clone();
                     }
                     if self.serial_devices.number_of_plots[self.device_idx] > 0 {
-
                         if self.data.dataset.len() != self.labels.len() {
                             self.labels = (0..max(self.data.dataset.len(), 1))
                                 .map(|i| format!("Column {i}"))
@@ -372,11 +285,13 @@ impl MyApp {
                             }
                         }
 
-                        let t_fmt =
-                            |x: GridMark, _range: &RangeInclusive<f64>| format!("{:4.2} s", x.value);
+                        let t_fmt = |x: GridMark, _range: &RangeInclusive<f64>| {
+                            format!("{:4.2} s", x.value)
+                        };
 
                         let plots_ui = ui.vertical(|ui| {
-                            for graph_idx in 0..self.serial_devices.number_of_plots[self.device_idx] {
+                            for graph_idx in 0..self.serial_devices.number_of_plots[self.device_idx]
+                            {
                                 if graph_idx != 0 {
                                     ui.separator();
                                 }
@@ -429,10 +344,8 @@ impl MyApp {
                         plot_ui_heigh = 0.0;
                     }
 
-                    let serial_height = panel_height
-                        - plot_ui_heigh
-                        - left_border * 2.0
-                        - top_spacing;
+                    let serial_height =
+                        panel_height - plot_ui_heigh - left_border * 2.0 - top_spacing;
 
                     let num_rows = self.data.raw_traffic.len();
                     let row_height = ui.text_style_height(&egui::TextStyle::Body);
@@ -466,10 +379,13 @@ impl MyApp {
                                 .collect();
 
                             let mut layouter = |ui: &egui::Ui, string: &str, wrap_width: f32| {
-                                let mut layout_job = highlight_impl(ui.ctx(),
-                                                                string,
-                                                                self.serial_devices.highlight_labels[self.device_idx].clone(),
-                                                                Color32::from_rgb(155, 164, 167)).unwrap();
+                                let mut layout_job = highlight_impl(
+                                    ui.ctx(),
+                                    string,
+                                    self.serial_devices.highlight_labels[self.device_idx].clone(),
+                                    Color32::from_rgb(155, 164, 167),
+                                )
+                                .unwrap();
                                 layout_job.wrap.max_width = wrap_width;
                                 ui.fonts(|f| f.layout_job(layout_job))
                             };
@@ -480,7 +396,7 @@ impl MyApp {
                                     .lock_focus(true)
                                     .text_color(color)
                                     .desired_width(width)
-                                    .layouter(&mut layouter)
+                                    .layouter(&mut layouter),
                             );
                         });
                     ui.horizontal(|ui| {
@@ -498,10 +414,7 @@ impl MyApp {
                             self.index = self.history.len() - 1;
                             let eol = self.eol.replace("\\r", "\r").replace("\\n", "\n");
                             if let Err(err) = self.send_tx.send(self.command.clone() + &eol) {
-                                print_to_console(
-                                    &self.print_lock,
-                                    Print::Error(format!("send_tx thread send failed: {:?}", err)),
-                                );
+                                log::error!("send_tx thread send failed: {:?}", err);
                             }
                             // stay in focus!
                             cmd_line.request_focus();
@@ -526,8 +439,7 @@ impl MyApp {
         });
     }
 
-    fn draw_serial_settings(&mut self, ctx: &egui::Context, ui :&mut Ui)
-    {
+    fn draw_serial_settings(&mut self, ctx: &egui::Context, ui: &mut Ui) {
         ui.horizontal(|ui| {
             ui.heading("Serial Monitor");
             self.paint_connection_indicator(ui);
@@ -575,7 +487,8 @@ impl MyApp {
                                 let dev_text = dev.replace("/dev/tty.", "");
                                 ui.selectable_value(&mut self.device, dev, dev_text);
                             });
-                    }).response;
+                    })
+                    .response;
                 // let selected_new_device = response.changed();  //somehow this does not work
                 // if selected_new_device {
                 if old_name != self.device {
@@ -611,12 +524,18 @@ impl MyApp {
                         self.serial_devices.devices.push(device);
                         self.serial_devices.number_of_plots.push(1);
                         self.serial_devices.number_of_highlights.push(1);
-                        self.serial_devices.highlight_labels.push(vec!["".to_string()]);
-                        self.serial_devices.labels.push(vec!["Column 0".to_string()]);
+                        self.serial_devices
+                            .highlight_labels
+                            .push(vec!["".to_string()]);
+                        self.serial_devices
+                            .labels
+                            .push(vec!["Column 0".to_string()]);
                         self.device_idx = self.serial_devices.devices.len() - 1;
                         save_serial_settings(&self.serial_devices);
                     }
-                    self.clear_tx.send(true).expect("failed to send clear after choosing new device");
+                    self.clear_tx
+                        .send(true)
+                        .expect("failed to send clear after choosing new device");
                     // need to clear the data here such that we don't get errors in the gui (plot)
                     self.data = DataContainer::default();
                     self.show_warning_window = WindowFeedback::None;
@@ -627,7 +546,10 @@ impl MyApp {
                 }
             }
             egui::ComboBox::from_id_salt("Baud Rate")
-                .selected_text(format!("{}", self.serial_devices.devices[self.device_idx].baud_rate))
+                .selected_text(format!(
+                    "{}",
+                    self.serial_devices.devices[self.device_idx].baud_rate
+                ))
                 .width(80.0)
                 .show_ui(ui, |ui| {
                     if self.connected_to_device {
@@ -641,7 +563,11 @@ impl MyApp {
                         );
                     });
                 });
-            let connect_text = if self.connected_to_device { "Disconnect" } else { "Connect" };
+            let connect_text = if self.connected_to_device {
+                "Disconnect"
+            } else {
+                "Connect"
+            };
             if ui.button(connect_text).clicked() {
                 if let Ok(mut device) = self.device_lock.write() {
                     if self.connected_to_device {
@@ -668,120 +594,184 @@ impl MyApp {
                 ui.disable();
             }
             egui::ComboBox::from_id_salt("Data Bits")
-                .selected_text(self.serial_devices.devices[self.device_idx].data_bits.to_string())
+                .selected_text(
+                    self.serial_devices.devices[self.device_idx]
+                        .data_bits
+                        .to_string(),
+                )
                 .width(30.0)
                 .show_ui(ui, |ui| {
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].data_bits, DataBits::Eight, DataBits::Eight.to_string());
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].data_bits, DataBits::Seven, DataBits::Seven.to_string());
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].data_bits, DataBits::Six, DataBits::Six.to_string());
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].data_bits, DataBits::Five, DataBits::Five.to_string());
-
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].data_bits,
+                        DataBits::Eight,
+                        DataBits::Eight.to_string(),
+                    );
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].data_bits,
+                        DataBits::Seven,
+                        DataBits::Seven.to_string(),
+                    );
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].data_bits,
+                        DataBits::Six,
+                        DataBits::Six.to_string(),
+                    );
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].data_bits,
+                        DataBits::Five,
+                        DataBits::Five.to_string(),
+                    );
                 });
             egui::ComboBox::from_id_salt("Parity")
-                .selected_text(self.serial_devices.devices[self.device_idx].parity.to_string())
+                .selected_text(
+                    self.serial_devices.devices[self.device_idx]
+                        .parity
+                        .to_string(),
+                )
                 .width(30.0)
                 .show_ui(ui, |ui| {
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].parity, Parity::None, Parity::None.to_string());
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].parity, Parity::Odd, Parity::Odd.to_string());
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].parity, Parity::Even, Parity::Even.to_string());
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].parity,
+                        Parity::None,
+                        Parity::None.to_string(),
+                    );
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].parity,
+                        Parity::Odd,
+                        Parity::Odd.to_string(),
+                    );
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].parity,
+                        Parity::Even,
+                        Parity::Even.to_string(),
+                    );
                 });
             egui::ComboBox::from_id_salt("Stop Bits")
-                .selected_text(self.serial_devices.devices[self.device_idx].stop_bits.to_string())
+                .selected_text(
+                    self.serial_devices.devices[self.device_idx]
+                        .stop_bits
+                        .to_string(),
+                )
                 .width(30.0)
                 .show_ui(ui, |ui| {
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].stop_bits, StopBits::One, StopBits::One.to_string());
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].stop_bits, StopBits::Two, StopBits::Two.to_string());
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].stop_bits,
+                        StopBits::One,
+                        StopBits::One.to_string(),
+                    );
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].stop_bits,
+                        StopBits::Two,
+                        StopBits::Two.to_string(),
+                    );
                 });
             egui::ComboBox::from_id_salt("Flow Control")
-                .selected_text(self.serial_devices.devices[self.device_idx].flow_control.to_string())
+                .selected_text(
+                    self.serial_devices.devices[self.device_idx]
+                        .flow_control
+                        .to_string(),
+                )
                 .width(75.0)
                 .show_ui(ui, |ui| {
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].flow_control, FlowControl::None, FlowControl::None.to_string());
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].flow_control, FlowControl::Hardware, FlowControl::Hardware.to_string());
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].flow_control, FlowControl::Software, FlowControl::Software.to_string());
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].flow_control,
+                        FlowControl::None,
+                        FlowControl::None.to_string(),
+                    );
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].flow_control,
+                        FlowControl::Hardware,
+                        FlowControl::Hardware.to_string(),
+                    );
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].flow_control,
+                        FlowControl::Software,
+                        FlowControl::Software.to_string(),
+                    );
                 });
             egui::ComboBox::from_id_salt("Timeout")
-                .selected_text(self.serial_devices.devices[self.device_idx].timeout.as_millis().to_string())
+                .selected_text(
+                    self.serial_devices.devices[self.device_idx]
+                        .timeout
+                        .as_millis()
+                        .to_string(),
+                )
                 .width(55.0)
                 .show_ui(ui, |ui| {
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].timeout, Duration::from_millis(0), "0");
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].timeout, Duration::from_millis(10), "10");
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].timeout, Duration::from_millis(100), "100");
-                    ui.selectable_value(&mut self.serial_devices.devices[self.device_idx].timeout, Duration::from_millis(1000), "1000");
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].timeout,
+                        Duration::from_millis(0),
+                        "0",
+                    );
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].timeout,
+                        Duration::from_millis(10),
+                        "10",
+                    );
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].timeout,
+                        Duration::from_millis(100),
+                        "100",
+                    );
+                    ui.selectable_value(
+                        &mut self.serial_devices.devices[self.device_idx].timeout,
+                        Duration::from_millis(1000),
+                        "1000",
+                    );
                 });
         });
     }
     fn draw_export_settings(&mut self, ctx: &egui::Context, ui: &mut Ui) {
         egui::Grid::new("export_settings")
-        .num_columns(2)
-        .spacing(Vec2 { x: 10.0, y: 10.0 })
-        .striped(true)
-        .show(ui, |ui| {
-            if ui.button(egui::RichText::new(format!("{} Save CSV", egui_phosphor::regular::FLOPPY_DISK)))
-                .on_hover_text("Save Plot Data to CSV.")
-                .clicked() || ui.input_mut(|i| i.consume_shortcut(&SAVE_FILE_SHORTCUT))
-            {
-                if let Some(path) = rfd::FileDialog::new().save_file() {
-                    self.picked_path = path;
-                    self.picked_path.set_extension("csv");
-                    if let Err(e) = self.save_tx.send(FileOptions {
-                        file_path: self.picked_path.clone(),
-                        save_absolute_time: self.gui_conf.save_absolute_time,
-                        save_raw_traffic: self.save_raw,
-                        names: self.serial_devices.labels[self.device_idx].clone(),
-                    }) {
-                        print_to_console(
-                            &self.print_lock,
-                            Print::Error(format!(
-                                "save_tx thread send failed: {:?}",
-                                e
-                            )),
-                        );
+            .num_columns(2)
+            .spacing(Vec2 { x: 10.0, y: 10.0 })
+            .striped(true)
+            .show(ui, |ui| {
+                if ui
+                    .button(egui::RichText::new(format!(
+                        "{} Save CSV",
+                        egui_phosphor::regular::FLOPPY_DISK
+                    )))
+                    .on_hover_text("Save Plot Data to CSV.")
+                    .clicked()
+                    || ui.input_mut(|i| i.consume_shortcut(&SAVE_FILE_SHORTCUT))
+                {
+                    if let Some(path) = rfd::FileDialog::new().save_file() {
+                        self.picked_path = path;
+                        self.picked_path.set_extension("csv");
+                        if let Err(e) = self.save_tx.send(FileOptions {
+                            file_path: self.picked_path.clone(),
+                            save_absolute_time: self.gui_conf.save_absolute_time,
+                            save_raw_traffic: self.save_raw,
+                            names: self.serial_devices.labels[self.device_idx].clone(),
+                        }) {
+                            log::error!("save_tx thread send failed: {:?}", e);
+                        }
                     }
+                };
+
+                if ui
+                    .button(egui::RichText::new(format!(
+                        "{} Save Plot",
+                        egui_phosphor::regular::FLOPPY_DISK
+                    )))
+                    .on_hover_text("Save an image of the Plot.")
+                    .clicked()
+                    || ui.input_mut(|i| i.consume_shortcut(&SAVE_PLOT_SHORTCUT))
+                {
+                    ctx.send_viewport_cmd(egui::ViewportCommand::Screenshot);
                 }
-            };
-
-            if ui
-                .button(egui::RichText::new(format!("{} Save Plot", egui_phosphor::regular::FLOPPY_DISK)))
-                .on_hover_text("Save an image of the Plot.")
-                .clicked() || ui.input_mut(|i| i.consume_shortcut(&SAVE_PLOT_SHORTCUT))
-
-            {
-                ctx.send_viewport_cmd(egui::ViewportCommand::Screenshot);
-            }
-            ui.end_row();
-            if ui.button(egui::RichText::new(format!("{} Clear Data", egui_phosphor::regular::X)))
-                .on_hover_text("Clear Data from Plot.")
-                .clicked() || ui.input_mut(|i| i.consume_shortcut(&CLEAR_PLOT_SHORTCUT)) {
-                print_to_console(
-                    &self.print_lock,
-                    Print::Ok("Cleared recorded Data".to_string()),
-                );
-                if let Err(err) = self.clear_tx.send(true) {
-                    print_to_console(
-                        &self.print_lock,
-                        Print::Error(format!(
-                            "clear_tx thread send failed: {:?}",
-                            err
-                        )),
-                    );
-                }
-                // need to clear the data here in order to prevent errors in the gui (plot)
-                self.data = DataContainer::default();
-                // self.names_tx.send(self.serial_devices.labels[self.device_idx].clone()).expect("Failed to send names");
-
-            }
-            ui.end_row();
-            ui.label("Save Raw Traffic");
-            ui.add(toggle(&mut self.save_raw))
-                .on_hover_text("Save second CSV containing raw traffic.")
-                .changed();
-            ui.end_row();
-            ui.label("Save Absolute Time");
-            ui.add(toggle(&mut self.gui_conf.save_absolute_time))
-                .on_hover_text("Save absolute time in CSV.");
-            ui.end_row();
-        });
+                ui.end_row();
+                ui.label("Save Raw Traffic");
+                ui.add(toggle(&mut self.save_raw))
+                    .on_hover_text("Save second CSV containing raw traffic.")
+                    .changed();
+                ui.end_row();
+                ui.label("Save Absolute Time");
+                ui.add(toggle(&mut self.gui_conf.save_absolute_time))
+                    .on_hover_text("Save absolute time in CSV.");
+                ui.end_row();
+            });
     }
 
     fn draw_global_settings(&mut self, ui: &mut Ui) {
@@ -804,19 +794,13 @@ impl MyApp {
             .clicked()
             || ui.input_mut(|i| i.consume_shortcut(&CLEAR_PLOT_SHORTCUT))
         {
-            print_to_console(
-                &self.print_lock,
-                Print::Ok("Cleared recorded Data".to_string()),
-            );
+            log::info!("Cleared recorded Data");
             if let Err(err) = self.clear_tx.send(true) {
-                print_to_console(
-                    &self.print_lock,
-                    Print::Error(format!("clear_tx thread send failed: {:?}", err)),
-                );
+                log::error!("clear_tx thread send failed: {:?}", err);
             }
             // need to clear the data here in order to prevent errors in the gui (plot)
             self.data = DataContainer::default();
-            //self.names_tx.send(self.serial_devices.plot_labels[self.device_idx].clone()).expect("Failed to send names");
+            // self.names_tx.send(self.serial_devices.labels[self.device_idx].clone()).expect("Failed to send names");
         }
         ui.horizontal(|ui| {
             if ui.button("Clear Device History").clicked() {
@@ -829,6 +813,19 @@ impl MyApp {
                 // self.serial_devices.labels[self.device_idx] = self.serial_devices.labels.clone();
             }
         });
+
+        ui.end_row();
+        ui.label("Show Sent Commands");
+        ui.add(toggle(&mut self.show_sent_cmds))
+            .on_hover_text("Show sent commands in console.");
+        ui.end_row();
+        ui.label("Show Timestamp");
+        ui.add(toggle(&mut self.show_timestamps))
+            .on_hover_text("Show timestamp in console.");
+        ui.end_row();
+        ui.label("EOL character");
+        ui.add(egui::TextEdit::singleline(&mut self.eol).desired_width(ui.available_width() * 0.9))
+            .on_hover_text("Configure your EOL character for sent commands..");
     }
 
     fn draw_plot_settings(&mut self, ui: &mut Ui) {
@@ -898,23 +895,6 @@ impl MyApp {
                             (self.serial_devices.number_of_plots[self.device_idx] + 1).clamp(0, 10);
                     }
                 });
-
-                ui.end_row();
-                ui.label("Show Sent Commands");
-                ui.add(toggle(&mut self.show_sent_cmds))
-                    .on_hover_text("Show sent commands in console.");
-                ui.end_row();
-                ui.label("Show Timestamp");
-                ui.add(toggle(&mut self.show_timestamps))
-                    .on_hover_text("Show timestamp in console.");
-                ui.end_row();
-                ui.label("EOL character");
-                ui.add(
-                    egui::TextEdit::singleline(&mut self.eol)
-                        .desired_width(ui.available_width() * 0.9),
-                )
-                .on_hover_text("Configure your EOL character for sent commands..");
-                // ui.checkbox(&mut self.gui_conf.debug, "Debug Mode");
                 ui.end_row();
             });
         ui.add_space(25.0);
@@ -1032,9 +1012,12 @@ impl MyApp {
         ui.add_space(5.0);
         for i in 0..(self.serial_devices.number_of_highlights[self.device_idx]) {
             ui.add(
-                egui::TextEdit::singleline(&mut self.serial_devices.highlight_labels[self.device_idx][i])
-                    .desired_width(0.95 * RIGHT_PANEL_WIDTH)
-            ).on_hover_text("Sentence to highlight");
+                egui::TextEdit::singleline(
+                    &mut self.serial_devices.highlight_labels[self.device_idx][i],
+                )
+                .desired_width(0.95 * RIGHT_PANEL_WIDTH),
+            )
+            .on_hover_text("Sentence to highlight");
             /*
             // Todo implement the color picker for each sentence
             let mut theme =
@@ -1056,63 +1039,35 @@ impl MyApp {
             .show(ctx, |ui| {
                 egui::ScrollArea::vertical().show(ui, |ui| {
                     ui.add_enabled_ui(true, |ui| {
-
-                        self.draw_serial_settings(ctx,ui);
+                        self.draw_serial_settings(ctx, ui);
 
                         self.draw_global_settings(ui);
                         ui.add_space(10.0);
-                        CollapsingHeader::new("Plot")
+                        CollapsingHeader::new("Plot Settings")
                             .default_open(true)
                             .show(ui, |ui| {
                                 self.draw_plot_settings(ui);
                             });
 
-                        CollapsingHeader::new("Text Highlight")
+                        CollapsingHeader::new("Text Highlight Settings")
                             .default_open(true)
                             .show(ui, |ui| {
                                 self.draw_highlight_settings(ctx, ui);
                             });
 
-                        CollapsingHeader::new("Export")
+                        CollapsingHeader::new("Export Settings")
                             .default_open(true)
                             .show(ui, |ui| {
                                 self.draw_export_settings(ctx, ui);
                             });
-
-                        if let Ok(read_guard) = self.print_lock.read() {
-                            self.console = read_guard.clone();
-                        }
-                        let num_rows = self.console.len();
-                        let row_height = ui.text_style_height(&egui::TextStyle::Body);
-
-                        ui.add_space(20.0);
-                        ui.separator();
-                        ui.label("Debug Info:");
-                        ui.add_space(5.0);
-                        egui::ScrollArea::vertical()
-                            .id_salt("console_scroll_area")
-                            .auto_shrink([false; 2])
-                            .stick_to_bottom(true)
-                            .max_height(row_height * 15.5)
-                            .show_rows(ui, row_height, num_rows, |ui, _row_range| {
-                                let content: String = self
-                                    .console
-                                    .iter()
-                                    .flat_map(|row| row.scroll_area_message(&self.gui_conf))
-                                    .map(|msg| msg.label + msg.content.as_str())
-                                    .collect::<Vec<_>>()
-                                    .join("\n");
-                                // we need to add it as one multiline object, such that we can select and copy
-                                // text over multiple lines
-                                ui.add(
-                                    egui::TextEdit::multiline(&mut content.as_str())
-                                        .font(DEFAULT_FONT_ID) // for cursor height
-                                        .lock_focus(true), // TODO: add a layouter to highlight the labels
-                                );
-                            });
                     });
+                    ui.add_space(20.0);
+                    ui.separator();
+                    ui.collapsing("Debug logs:", |ui| {
+                        egui_logger::logger_ui().show(ui);
+                    });
+                });
             });
-        });
     }
 
     fn paint_connection_indicator(&self, ui: &mut egui::Ui) {
@@ -1172,7 +1127,7 @@ impl eframe::App for MyApp {
                     image::ColorType::Rgba8,
                 )
                 .unwrap();
-                eprintln!("Image saved to {path:?}.");
+                println!("Image saved to {path:?}.");
             }
         }
     }